+++ index.ts
@@ -1,6 +1,7 @@
 import {IPersonData} from './src/interfaces/person.interface';
 import {Person} from './src/person/person.class';
 import {PersonCollection} from './src/person/person.collection';
+import {UniqueCollection} from './src/collection/unique-collection.class';
 
 const personsData: IPersonData[] = [
     {name: 'Vladlen', age: 25},
@@ -13,18 +14,43 @@ const personsData: IPersonData[] = [
 ];
 
 const persons: Person[] = personsData.map((people: IPersonData): Person => new Person(people));
-const pc: PersonCollection = new PersonCollection(persons);
 
-// Показать имена и возвраст всех, затем только имена для 18+
+// Отсортировать по возврасту, показать имена и возвраст всех, затем только имена для 18+
+const pc: PersonCollection = new PersonCollection(persons);
 pc
-.call((people: Person) => { console.log(`${people.getName()} - ${people.age}`)} )
-.filter((people: Person) => people.age >= 18)
-.showName();
+.sort((a: Person, b: Person) => a.age - b.age)
+.call((person: Person) => { console.log(`${person.getName()} - ${person.age}`)} )
+.filter((person: Person) => person.age >= 18)
+.showName()
+.do(() => console.log('<===================================>'));
 
 // Тоже самое, без урезания коллекции
-// pc
-// .call((people: Person) => { console.log(`${people.getName()} - ${people.age}`)} )
+// const pc1: PersonCollection = new PersonCollection(persons);
+// pc1
+// .sort((a: Person, b: Person) => a.age - b.age)
+// .call((person: Person) => { console.log(`${person.getName()} - ${person.age}`)} )
 // .call({
-//     filter: ((people: Person) => people.age >= 18),
-//     call: (people: Person) => { people.showName(); },
-// });
+//     filter: ((person: Person) => person.age >= 18),
+//     call: (person: Person) => { person.showName(); },
+// })
+// .do(() => console.log('<===================================>'));
+
+// Тоже самое через цикл for..of (без сортировки)
+// const pc2: PersonCollection = new PersonCollection(persons);
+// for (const person of pc2) {
+//     console.log(`${person.getName()} - ${person.age}`);
+// }
+// for (const person of pc2) {
+//     if (person.age >= 18) {
+//         person.showName();
+//     }
+// }
+// console.log('<===================================>')
+
+// Уникальная коллекция
+const uc: UniqueCollection<Person> = new UniqueCollection(persons);
+uc
+.rpush(persons)
+.rpush(persons)
+.call((person: Person) => { person.showName(); })
+.do(() => console.log('<===================================>'));
